fix: refactor mock directory
