+++ src/mock/mock-utils.js
@@ -1,0 +1,13 @@
+import {PHOTO} from './mock-const';
+
+
+function getRandomArrayElement(items) {
+  return items[Math.floor(Math.random() * items.length)];
+}
+
+function getRandomPhotoURL() {
+  const randomNumber = Math.floor(Math.random() * 10) + 1;
+  return `${PHOTO}${randomNumber}`;
+}
+
+export {getRandomArrayElement, getRandomPhotoURL};
