+++ src/mock/points.js
@@ -1,92 +1,7 @@
-import {getRandomArrayElement, getRandomPhotoURL} from '../utils.js';
-import {FILTERS, SORT, POINT_TYPE, CITIES, OFFERS, DESCRIPTION, PRICES} from '../const.js';
-
-const mockDestination = [
-  {
-    id: '1',
-    description: getRandomArrayElement(DESCRIPTION),
-    name: getRandomArrayElement(CITIES),
-    pictures: [
-      {
-        src: getRandomPhotoURL(),
-        description: 'description №1'
-      }
-    ]
-  },
-  {
-    id: '2',
-    description: getRandomArrayElement(DESCRIPTION),
-    name: getRandomArrayElement(CITIES),
-    pictures: [
-      {
-        src: getRandomPhotoURL(),
-        description: 'description №2'
-      },
-      {
-        src: getRandomPhotoURL(),
-        description: 'description №3'
-      }
-    ]
-  },
-  {
-    id: '3',
-    description: getRandomArrayElement(DESCRIPTION),
-    name: getRandomArrayElement(CITIES),
-    pictures: [
-      {
-        src: getRandomPhotoURL(),
-        description: 'description №4'
-      },
-      {
-        src: getRandomPhotoURL(),
-        description: 'description №5'
-      },
-      {
-        src: getRandomPhotoURL(),
-        description: 'description №6'
-      }
-    ]
-  }
-];
-
-const mockOffers = [
-  {
-    type: getRandomArrayElement(POINT_TYPE),
-    offers: [
-      {
-        id: 'id_1',
-        title: getRandomArrayElement(OFFERS),
-        price: getRandomArrayElement(PRICES)
-      }
-    ]
-  },
-  {
-    type: getRandomArrayElement(POINT_TYPE),
-    offers: [
-      {
-        id: 'id_2',
-        title: getRandomArrayElement(OFFERS),
-        price: getRandomArrayElement(PRICES)
-      },
-      {
-        id: 'id_3',
-        title: getRandomArrayElement(OFFERS),
-        price: getRandomArrayElement(PRICES)
-      }
-    ]
-  },
-  {
-    type: getRandomArrayElement(POINT_TYPE),
-    offers: [
-      {
-        id: 'id_4',
-        title: getRandomArrayElement(OFFERS),
-        price: getRandomArrayElement(PRICES)
-      }
-    ]
-  }
-];
-
+import {getRandomArrayElement} from './mock-utils.js';
+import {PRICES} from './mock-const.js';
+import {mockDestinations} from './destinations.js';
+import {mockOffers} from './offers.js';
 
 const mockPoints = [
   {
@@ -94,7 +9,7 @@ const mockPoints = [
     basePrice: getRandomArrayElement(PRICES),
     dateFrom: '2019-07-10T09:15:56.845Z',
     dateTo: '2019-07-11T04:55:13.375Z',
-    destination: mockDestination[0],
+    destination: mockDestinations[0],
     isFavorite: false,
     offers: [
       mockOffers[0].offers[0]
@@ -106,7 +21,7 @@ const mockPoints = [
     basePrice: getRandomArrayElement(PRICES),
     dateFrom: '2019-07-10T09:15:56.845Z',
     dateTo: '2019-07-11T04:55:13.375Z',
-    destination: mockDestination[1],
+    destination: mockDestinations[1],
     isFavorite: true,
     offers: [
       mockOffers[1].offers[0],
@@ -119,7 +34,7 @@ const mockPoints = [
     basePrice: getRandomArrayElement(PRICES),
     dateFrom: '2019-07-10T09:15:56.845Z',
     dateTo: '2019-07-11T04:55:13.375Z',
-    destination: mockDestination[2],
+    destination: mockDestinations[2],
     isFavorite: false,
     offers: [
       mockOffers[2].offers[0]
