+++ src/model/points-model.js
@@ -1,10 +1,22 @@
 import {getRandomPoint} from '../mock/points';
-import {POINT_COUNT} from '../const.js';
+import {POINT_COUNT} from '../mock/mock-const';
+import {mockOffers} from '../mock/offers';
+import {mockDestinations} from '../mock/destinations';
 
 export default class PointsModel {
   points = Array.from({length: POINT_COUNT}, getRandomPoint);
+  offers = mockOffers;
+  destinations = mockDestinations;
 
   getPoints() {
     return this.points;
+  }
+
+  getOffers() {
+    return this.offers;
+  }
+
+  getDestinations() {
+    return this.destinations;
   }
 }
